+++ cypress/e2e/invite/invite-dialog.spec.ts
@@ -166,8 +166,8 @@ describe("Invite dialog", function () {
 
         // Assert that the hovered user name on invitation UI does not have background color
         // TODO: implement the test on room-header.spec.ts
-        cy.get(".mx_RoomHeader").within(() => {
-            cy.get(".mx_RoomHeader_name--textonly")
+        cy.get(".mx_LegacyRoomHeader").within(() => {
+            cy.get(".mx_LegacyRoomHeader_name--textonly")
                 .realHover()
                 .should("have.css", "background-color", "rgba(0, 0, 0, 0)");
         });
