+++ res/css/structures/_MainSplit.pcss
@@ -27,7 +27,7 @@ limitations under the License.
     /* The resizer should be centered: only half of the gap-width is handled by the right panel. */
     /* The other half by the RoomView. */
     padding-left: calc(var(--container-gap-width) / 2);
-    height: calc(100vh - 51px); /* height of .mx_RoomHeader.light-panel */
+    height: calc(100vh - 51px); /* height of .mx_LegacyRoomHeader.light-panel */
 
     &:hover .mx_ResizeHandle--horizontal::before {
         position: absolute;
