+++ res/css/structures/_RoomView.pcss
@@ -190,7 +190,7 @@ limitations under the License.
 }
 
 /* Rooms with immersive content */
-.mx_RoomView_immersive .mx_RoomHeader_wrapper {
+.mx_RoomView_immersive .mx_LegacyRoomHeader_wrapper {
     border: unset;
 }
