+++ src/i18n/strings/en_EN.json
@@ -1009,6 +1009,7 @@
     "Hide notification dot (only display counters badges)": "Hide notification dot (only display counters badges)",
     "Enable intentional mentions": "Enable intentional mentions",
     "Enable ask to join": "Enable ask to join",
+    "Under active development, new room header & details interface": "Under active development, new room header & details interface",
     "Use a more compact 'Modern' layout": "Use a more compact 'Modern' layout",
     "Show a placeholder for removed messages": "Show a placeholder for removed messages",
     "Show join/leave messages (invites/removes/bans unaffected)": "Show join/leave messages (invites/removes/bans unaffected)",
@@ -1941,6 +1942,26 @@
     "Encrypted messages before this point are unavailable.": "Encrypted messages before this point are unavailable.",
     "You can't see earlier messages": "You can't see earlier messages",
     "Scroll to most recent messages": "Scroll to most recent messages",
+    "Video call (Jitsi)": "Video call (Jitsi)",
+    "Video call (%(brand)s)": "Video call (%(brand)s)",
+    "Ongoing call": "Ongoing call",
+    "You do not have permission to start video calls": "You do not have permission to start video calls",
+    "There's no one here to call": "There's no one here to call",
+    "You do not have permission to start voice calls": "You do not have permission to start voice calls",
+    "Freedom": "Freedom",
+    "Spotlight": "Spotlight",
+    "Change layout": "Change layout",
+    "Forget room": "Forget room",
+    "Hide Widgets": "Hide Widgets",
+    "Show Widgets": "Show Widgets",
+    "Search": "Search",
+    "Close call": "Close call",
+    "View chat timeline": "View chat timeline",
+    "Room options": "Room options",
+    "Join Room": "Join Room",
+    "(~%(count)s results)|other": "(~%(count)s results)",
+    "(~%(count)s results)|one": "(~%(count)s result)",
+    "Video rooms are a beta feature": "Video rooms are a beta feature",
     "Show %(count)s other previews|other": "Show %(count)s other previews",
     "Show %(count)s other previews|one": "Show %(count)s other preview",
     "Close preview": "Close preview",
@@ -2018,26 +2039,6 @@
     "Room %(name)s": "Room %(name)s",
     "Recently visited rooms": "Recently visited rooms",
     "No recently visited rooms": "No recently visited rooms",
-    "Video call (Jitsi)": "Video call (Jitsi)",
-    "Video call (%(brand)s)": "Video call (%(brand)s)",
-    "Ongoing call": "Ongoing call",
-    "You do not have permission to start video calls": "You do not have permission to start video calls",
-    "There's no one here to call": "There's no one here to call",
-    "You do not have permission to start voice calls": "You do not have permission to start voice calls",
-    "Freedom": "Freedom",
-    "Spotlight": "Spotlight",
-    "Change layout": "Change layout",
-    "Forget room": "Forget room",
-    "Hide Widgets": "Hide Widgets",
-    "Show Widgets": "Show Widgets",
-    "Search": "Search",
-    "Close call": "Close call",
-    "View chat timeline": "View chat timeline",
-    "Room options": "Room options",
-    "Join Room": "Join Room",
-    "(~%(count)s results)|other": "(~%(count)s results)",
-    "(~%(count)s results)|one": "(~%(count)s result)",
-    "Video rooms are a beta feature": "Video rooms are a beta feature",
     "Video room": "Video room",
     "Public space": "Public space",
     "Public room": "Public room",
@@ -2242,11 +2243,11 @@
     "Yours, or the other users' session": "Yours, or the other users' session",
     "Error starting verification": "Error starting verification",
     "We were unable to start a chat with the other user.": "We were unable to start a chat with the other user.",
-    "Nothing pinned, yet": "Nothing pinned, yet",
-    "If you have permissions, open the menu on any message and select <b>Pin</b> to stick them here.": "If you have permissions, open the menu on any message and select <b>Pin</b> to stick them here.",
     "Pinned messages": "Pinned messages",
     "Chat": "Chat",
     "Room info": "Room info",
+    "Nothing pinned, yet": "Nothing pinned, yet",
+    "If you have permissions, open the menu on any message and select <b>Pin</b> to stick them here.": "If you have permissions, open the menu on any message and select <b>Pin</b> to stick them here.",
     "You can only pin up to %(count)s widgets|other": "You can only pin up to %(count)s widgets",
     "Maximise": "Maximise",
     "Unpin this widget to view it in this panel": "Unpin this widget to view it in this panel",
