+++ src/utils/exportUtils/HtmlExport.tsx
@@ -181,23 +181,23 @@ export default class HTMLExporter extends Exporter {
                 <div class="mx_MatrixChat_wrapper" aria-hidden="false">
                     <div class="mx_MatrixChat">
                     <main class="mx_RoomView">
-                        <div class="mx_RoomHeader light-panel">
-                        <div class="mx_RoomHeader_wrapper" aria-owns="mx_RightPanel">
-                            <div class="mx_RoomHeader_avatar">
+                        <div class="mx_LegacyRoomHeader light-panel">
+                        <div class="mx_LegacyRoomHeader_wrapper" aria-owns="mx_RightPanel">
+                            <div class="mx_LegacyRoomHeader_avatar">
                             <div class="mx_DecoratedRoomAvatar">
                                ${roomAvatar}
                             </div>
                             </div>
-                            <div class="mx_RoomHeader_name">
+                            <div class="mx_LegacyRoomHeader_name">
                             <div
                                 dir="auto"
-                                class="mx_RoomHeader_nametext"
+                                class="mx_LegacyRoomHeader_nametext"
                                 title="${safeRoomName}"
                             >
                                 ${safeRoomName}
                             </div>
                             </div>
-                            <div class="mx_RoomHeader_topic" dir="auto"> ${safeTopic} </div>
+                            <div class="mx_LegacyRoomHeader_topic" dir="auto"> ${safeTopic} </div>
                         </div>
                         </div>
                         ${previousMessagesLink}
